# In tests use a smaller version of `matches_mapper` so they run faster