--- a/tests/testthat/helper.R
+++ b/tests/testthat/helper.R
@@ -1,0 +1,1 @@
+small_matches_mapper <- head(tiltIndicatorAfter::matches_mapper, 100)