--- a/tests/testthat/test-prepare_inter_pctr_product.R
+++ b/tests/testthat/test-prepare_inter_pctr_product.R
@@ -1,5 +1,5 @@
 test_that("`avg_matching_certainty` should have only one value", {
-  out <- prepare_inter_pctr_product(pctr_product, ep_companies, ecoinvent_activities, matches_mapper, isic_tilt_mapper) |>
+  out <- prepare_inter_pctr_product(pctr_product, ep_companies, ecoinvent_activities, small_matches_mapper, isic_tilt_mapper) |>
     group_by(companies_id) |>
     summarise(count = dplyr::n_distinct(avg_matching_certainty))
   expect_equal(unique(out$count), 1)