--- a/tests/testthat/test-prepare_inter_pstr_product.R
+++ b/tests/testthat/test-prepare_inter_pstr_product.R
@@ -1,5 +1,5 @@
 test_that("risk_category colummn should not have more than one NA for no result companies", {
-  out <- prepare_inter_pstr_product(pstr_product, ep_companies, ecoinvent_activities, matches_mapper) |>
+  out <- prepare_inter_pstr_product(pstr_product, ep_companies, ecoinvent_activities, small_matches_mapper) |>
     filter(is.na(risk_category)) |>
     group_by(companies_id) |>
     summarise(count = n())