--- a/tests/testthat/test-prepare_matches_mapper.R
+++ b/tests/testthat/test-prepare_matches_mapper.R
@@ -1,9 +1,9 @@
 test_that("`multi_match` column has either `TRUE` or `FALSE` values", {
-  out <- prepare_matches_mapper(matches_mapper, ecoinvent_activities)
+  out <- prepare_matches_mapper(small_matches_mapper, ecoinvent_activities)
   expect_equal(unique(out$multi_match), c("TRUE", "FALSE"))
 })
 
 test_that("`completion` column has either `low`, `medium`, or `high` values", {
-  out <- prepare_matches_mapper(matches_mapper, ecoinvent_activities)
+  out <- prepare_matches_mapper(small_matches_mapper, ecoinvent_activities)
   expect_equal(unique(out$completion), c("low", "high", "medium"))
 })