Closes #33 

This PR reduces the size of `matches_mapper` in tests only, so that they run faster. This reduces the tests run-time by an order of magnitude.

As it touches no production code I'll go ahead and merge without review.

NOW

```
ℹ Testing tiltIndicatorAfter
✔ | F W S  OK | Context
✔ |        13 | data [0.6s]                                                                       
✔ |         1 | prepare_ictr_company [0.1s]                                                       
✔ |         1 | prepare_inter_pctr_product                                                        
✔ |         1 | prepare_inter_pstr_product                                                        
✔ |         1 | prepare_istr_company                                                              
✔ |         2 | prepare_matches_mapper                                                            
✔ |         1 | prepare_pctr_company                                                              
✔ |         1 | prepare_pctr_product                                                              
✔ |         1 | prepare_pstr_company                                                              
✔ |         1 | prepare_pstr_product 
```

BEFORE

```
ℹ Testing tiltIndicatorAfter
✔ | F W S  OK | Context
✔ |        13 | data [0.6s]                                                     
✔ |         1 | prepare_ictr_company [1.7s]                                     
✔ |         1 | prepare_inter_pctr_product [1.8s]                               
✔ |         1 | prepare_inter_pstr_product [1.8s]                               
✔ |         1 | prepare_istr_company [1.9s]                                     
✔ |         2 | prepare_matches_mapper [3.4s]                                   
✔ |         1 | prepare_pctr_company [1.6s]                                     
✔ |         1 | prepare_pctr_product [1.8s]                                     
✔ |         1 | prepare_pstr_company [1.7s]                                     
✔ |         1 | prepare_pstr_product [1.9s]  
```